+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3
+
+* Backward incompatible changes for ShutdownErrorException, FatalErrorException and SandboxException
+(setters have been removed, constructor and parent class have been changed)
+
 ## 0.2
 
 * Limit for number of displayed steps in CLI